adding clean architecture
