+++ .dart_tool/flutter_build/d65bbed71c8b4421c94189f5fab168c8/debug_unpack_ios.stamp
@@ -1,0 +1,1 @@
+{"inputs":["/Users/vicklercharles/development/flutter/packages/flutter_tools/lib/src/build_system/targets/ios.dart","/Users/vicklercharles/development/flutter/bin/internal/engine.version"],"outputs":["/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/Flutter.framework/Flutter"]}
