+++ .dart_tool/flutter_build/d65bbed71c8b4421c94189f5fab168c8/gen_dart_plugin_registrant.stamp
@@ -1,0 +1,1 @@
+{"inputs":["/Users/vicklercharles/Desktop/flutter-portfolio-tracker/.dart_tool/package_config_subset"],"outputs":[]}
