+++ .dart_tool/flutter_build/d65bbed71c8b4421c94189f5fab168c8/outputs.json
@@ -1,0 +1,1 @@
+["/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/Flutter.framework/Flutter","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/vm_snapshot_data","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/isolate_snapshot_data","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/kernel_blob.bin","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/App","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/Info.plist","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-Black.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-BlackItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-Bold.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-BoldItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-ExtraBold.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-ExtraBoldItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-ExtraLight.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-ExtraLightItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-Italic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-Light.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-LightItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-Medium.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-MediumItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-Regular.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-SemiBold.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-SemiBoldItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-Thin.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/poppins/Poppins-ThinItalic.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/packages/cupertino_icons/assets/CupertinoIcons.ttf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/fonts/MaterialIcons-Regular.otf","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/shaders/ink_sparkle.frag","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/AssetManifest.json","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/FontManifest.json","/Users/vicklercharles/Desktop/flutter-portfolio-tracker/build/ios/Debug-iphonesimulator/App.framework/flutter_assets/NOTICES.Z"]
