+++ .dart_tool/package_config.json
@@ -164,6 +164,12 @@
       "languageVersion": "2.12"
     },
     {
+      "name": "flutter_riverpod",
+      "rootUri": "file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_riverpod-2.3.6",
+      "packageUri": "lib/",
+      "languageVersion": "2.17"
+    },
+    {
       "name": "flutter_test",
       "rootUri": "file:///Users/vicklercharles/development/flutter/packages/flutter_test",
       "packageUri": "lib/",
@@ -302,6 +308,12 @@
       "languageVersion": "2.18"
     },
     {
+      "name": "riverpod",
+      "rootUri": "file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/riverpod-2.3.6",
+      "packageUri": "lib/",
+      "languageVersion": "2.17"
+    },
+    {
       "name": "shelf",
       "rootUri": "file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/shelf-1.4.0",
       "packageUri": "lib/",
@@ -334,6 +346,12 @@
     {
       "name": "stack_trace",
       "rootUri": "file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0",
+      "packageUri": "lib/",
+      "languageVersion": "2.12"
+    },
+    {
+      "name": "state_notifier",
+      "rootUri": "file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/state_notifier-0.7.2+1",
       "packageUri": "lib/",
       "languageVersion": "2.12"
     },
@@ -410,7 +428,7 @@
       "languageVersion": "2.16"
     }
   ],
-  "generated": "2023-03-06T06:45:57.891727Z",
+  "generated": "2023-05-03T09:43:58.541759Z",
   "generator": "pub",
   "generatorVersion": "2.18.2"
 }
