+++ .dart_tool/package_config_subset
@@ -1,7 +1,7 @@
 assetdash_takehome
 2.16
-file:///Users/vicklercharles/Downloads/assetdash-mobile-takehome-main%202/
-file:///Users/vicklercharles/Downloads/assetdash-mobile-takehome-main%202/lib/
+file:///Users/vicklercharles/Desktop/flutter-portfolio-tracker/
+file:///Users/vicklercharles/Desktop/flutter-portfolio-tracker/lib/
 _fe_analyzer_shared
 2.17
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/_fe_analyzer_shared-50.0.0/
@@ -106,6 +106,10 @@ flutter_lints
 2.12
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/lib/
+flutter_riverpod
+2.17
+file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_riverpod-2.3.6/
+file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_riverpod-2.3.6/lib/
 frontend_server_client
 2.12
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/frontend_server_client-3.2.0/
@@ -194,6 +198,10 @@ pubspec_parse
 2.18
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/pubspec_parse-1.2.2/
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/pubspec_parse-1.2.2/lib/
+riverpod
+2.17
+file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/riverpod-2.3.6/
+file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/riverpod-2.3.6/lib/
 shelf
 2.17
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/shelf-1.4.0/
@@ -214,6 +222,10 @@ stack_trace
 2.12
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/lib/
+state_notifier
+2.12
+file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/state_notifier-0.7.2+1/
+file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/state_notifier-0.7.2+1/lib/
 stream_channel
 2.12
 file:///Users/vicklercharles/development/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/
