+++ README.md
@@ -2,8 +2,7 @@
 
 Flutter app that retrieves data from an API and displays it in a pie chart and a table/list view. The app also allow users to filter the table/list view based on the asset type.
 
-
-## How to Use 
+## How to Use
 
 Project Requirements:
 
@@ -21,20 +20,19 @@ https://github.com/vickcharles/assetdash-mobile-takehome-main.git
 
 **Step 2:**
 
-Go to project root and execute the following command in console to get the required dependencies: 
+Go to project root and execute the following command in console to get the required dependencies:
 
 ```
-flutter pub get 
+flutter pub get
 ```
 
 **Step 3:**
 
-Run the project 
+Run the project
 
 ```
 flutter run
 ```
-
 
 ## Runn the test
 
@@ -48,5 +46,3 @@ flutter test
 - [Mockito](https://pub.dev/packages/mockito)
 - [Pie Chart](https://pub.dev/packages/pie_chart)
 - [Build_runner](https://pub.dev/packages/pie_chart)
-
-
