--- assetdash_takehome/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: 097d3313d8e2c7f901932d63e537c1acefb87800
-  channel: stable
-
-project_type: app
