--- data/mock_api.json
@@ -1,76 +1,0 @@
-{
-  "holdings": [
-    {
-      "ticker": "AAPL",
-      "name": "Apple",
-      "type": "stock",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "TSLA",
-      "name": "Tesla",
-      "type": "stock",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "BTC",
-      "name": "Bitcoin",
-      "type": "crypto",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "ETH",
-      "name": "Ethereum",
-      "type": "crypto",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "10YR",
-      "name": "10 Year Bond",
-      "type": "bonds",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "20YR",
-      "name": "20 Year Bond",
-      "type": "bonds",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "BAYC",
-      "name": "Bored Ape Yacht Club",
-      "type": "nft",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "SMB",
-      "name": "Solana Monkey Business",
-      "type": "stock",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "JOE",
-      "name": "Trader Joe",
-      "type": "defi",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "OHM",
-      "name": "Olympus",
-      "type": "defi",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "FUNDRISE",
-      "name": "Fundrise",
-      "type": "real_estate",
-      "value": "$datatype.number"
-    },
-    {
-      "ticker": "CONCREIT",
-      "name": "Concreit",
-      "type": "real_estate",
-      "value": "$datatype.number"
-    }
-  ]
-}
