+++ lib/domain/repositories/holding_repository.dart
@@ -1,0 +1,5 @@
+import '../entities/holding.dart';
+
+abstract class HoldingRepository {
+  Future<List<Holding>> getHoldings({required int? userId});
+}
