+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   http: ^0.13.5
   mockito: ^5.3.2
   build_runner: ^2.3.3
+  flutter_riverpod: ^2.1.1
 
 dev_dependencies:
   flutter_test:
